Renamed MessageSender to UserMessageSender, created new MessageSender interface for sending messages without dependency on connection type
