+++ app/src/main/java/org/albard/dubito/app/messaging/MessageSender.java
@@ -1,0 +1,9 @@
+package org.albard.dubito.app.messaging;
+
+import java.net.InetSocketAddress;
+
+public interface MessageSender {
+    void sendToAll(Object message);
+
+    void sendTo(Object message, InetSocketAddress[] userEndPoints);
+}
